// lib/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.4"
  }
  public: {
    Tables: {
// ...
      users: {
        Row: {
          created_at: string
          deleted_at: string | null
          department_id: string | null
          email: string
          employee_id: string | null
          full_name: string
          id: string
          role: string
          status: string
          team_id: string | null
          updated_at: string
          username: string
        }
        Insert: {
          created_at?: string
          deleted_at?: string | null
          department_id?: string | null
          email: string
          employee_id?: string | null
          full_name: string
          id: string
          role?: string
          status?: string
          team_id?: string | null
          updated_at?: string
          username: string
        }
        Update: {
          created_at?: string
          deleted_at?: string | null
          department_id?: string | null
          email?: string
          employee_id?: string | null
          full_name?: string
          id?: string
          role?: string
          status?: string
          team_id?: string | null
          updated_at?: string
          username?: string
        }
        Relationships: [
          {
            foreignKeyName: "users_department_id_fkey"
            columns: ["department_id"]
            isOneToOne: false
            referencedRelation: "departments"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "users_team_id_fkey"
            columns: ["team_id"]
            isOneToOne: false
            referencedRelation: "teams"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
  }
}